refactor: ResultPage 컴포넌트에서 상태 관리 로직 제거 및 useUserInfo 훅 사용으로 변경
