+++ web/components/gacha/ResultPage.tsx
@@ -1,22 +1,17 @@
 'use client';
-import React, { useEffect, useState } from 'react';
 import Image from 'next/image';
 import Button from '@/components/common/Button';
 import { useRouter, useSearchParams } from 'next/navigation';
+import { useUserInfo } from '@/hooks/queries/useUserInfo';
+import { useCloverCount } from '@/hooks/queries/useCloverCount';
 
 function ResultPage() {
-  const [nickname, setNickName] = useState('');
-  const [clover, setClover] = useState(0);
   const searchParams = useSearchParams();
   const imageUrl = searchParams.get('url'); // url 파라미터 가져오기
   const id = searchParams.get('id'); // id 파라미터 가져오기
   const router = useRouter();
-  useEffect(() => {
-    setNickName(localStorage.getItem('nickname') || '');
-    setClover(Number(localStorage.getItem('cloverCount')) || 0);
-  }, []);
-  console.log('urlId:', id);
-
+  const { data: { nickname = '' } = {} } = useUserInfo();
+  const { data: clover } = useCloverCount();
   return (
     <div className="m-h-screen relative flex flex-grow flex-col items-center overflow-hidden">
       <p className="onboarding mt-25 text-center whitespace-pre-line">
