+++ web/hooks/queries/useUserInfo.ts
@@ -1,13 +1,8 @@
 import { useQuery } from '@tanstack/react-query';
 import { queryKeys } from '@/utils/queries/queryKeys';
 import { getUserInfo } from '@/utils/queries/member';
+import { UserInfo } from '@/types/userInfo';
 
-interface UserInfo {
-  nickname: string;
-  badgeUrl: string;
-  userId: string;
-  badgeId: number;
-}
 
 export const useUserInfo = () => {
   return useQuery<UserInfo>({
