+++ web/types/userInfo.ts
@@ -1,0 +1,8 @@
+interface UserInfo {
+  nickname: string;
+  badgeUrl: string;
+  userId: string;
+  badgeId: number;
+}
+
+export type { UserInfo };
